Restructured QUnit test runner
Except for the runTests.html file which may be used to run the tests manually and can be regarded
the test runner configuration file, all test runner files are located in the runner's lib directory
now. Consequently, Travis needs to run the PhantomJS runner within the lib directory pointing to the
HTML test runner per argument.